moving test case
git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@2958 67b5017d-e62d-0410-a871-f72b1c62810f

+++ Crough/HybridComp10.f
@@ -1,5 +1,5 @@
       program HybridComp10
-      integer i
+      integer*8 i
       integer j
       integer*8 a(5, 5)
       do 10 i = 1, 5
